refactor: hooks/common > hooks/@common으로 이동
