+++ frontend/src/components/Review/ReviewListAndChart/ReviewList/ReviewList.tsx
@@ -3,7 +3,7 @@ import { styled } from 'styled-components';
 
 import WriteIcon from '@/assets/svg/write_btn.svg';
 import { useValidParams } from '@/hooks/@common/useValidParams';
-import useValidQueryString from '@/hooks/common/useValidQueryString';
+import useValidQueryString from '@/hooks/@common/useValidQueryString';
 import { useReviewListQuery } from '@/hooks/query/review';
 import { routerPath } from '@/router/routes';
 import { zipgoLocalStorage } from '@/utils/localStorage';
