+++ frontend/src/hooks/@common/useImageUpload.ts
